Big model refactoring
