+++ Pirat/Model/Api/Resource/Offer.cs
