+++ Pirat/Model/Entity/Resource/Common/Deletable.cs
