+++ Pirat/Model/Entity/Resource/Common/Findable.cs
