+++ Pirat/Model/Entity/Resource/Common/Updatable.cs
