+++ Pirat/Model/Entity/Resource/Demand/DemandConsumableEntity.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Pirat.Model.EntityDemand
+{
+    public class DemandConsumableEntity
+    {
+    }
+}
