+++ Pirat/Model/Entity/Resource/Demand/DemandDeviceEntity.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Pirat.Model.Entity;
+
+namespace Pirat.Model.EntityDemand
+{
+
+}
