+++ Pirat/Model/Entity/Resource/RegionSubscription.cs
