+++ Pirat/Model/Entity/Resource/Stock/ChangeEntity.cs
