+++ Pirat/Model/Entity/Resource/Stock/ConsumableEntity.cs
