+++ Pirat/Model/Entity/Resource/Stock/DeviceEntity.cs
