+++ Pirat/Services/Resource/ResourceInputValidatorService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Pirat.Codes;
 using Pirat.Model;
+using Pirat.Model.Api.Resource.Stock;
 
 namespace Pirat.Services.Resource
 {
